Log file size on S3 upload
We have been having some problems with unexpected files being uploaded.
This change logs the path, size and acl of a saved file.
